[refactor] separate package
