+++ src/main/java/racing/domain/Car.java
@@ -1,4 +1,6 @@
-package racing;
+package racing.domain;
+
+import java.util.Objects;
 
 public class Car {
     private int position = 0;
@@ -7,8 +9,16 @@ public class Car {
     }
 
     public void move(CarMovement carMovement) {
+        this.validateCarMovement(carMovement);
+
         if (carMovement.isMove()) {
             this.position++;
+        }
+    }
+
+    private void validateCarMovement(CarMovement carMovement) {
+        if (Objects.isNull(carMovement)) {
+            throw new IllegalArgumentException();
         }
     }
