+++ src/main/java/racing/domain/CarMovement.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 @FunctionalInterface
 public interface CarMovement {
